move Formatter.ts to formatter/
