+++ src/languages/hive.formatter.ts
@@ -1,4 +1,4 @@
-import Formatter from 'src/core/Formatter';
+import Formatter from 'src/formatter/formatter';
 import Tokenizer from 'src/core/Tokenizer';
 import { dedupe } from 'src/utils';
