+++ src/languages/plsql.formatter.ts
@@ -1,4 +1,4 @@
-import Formatter from 'src/core/Formatter';
+import Formatter from 'src/formatter/formatter';
 import Tokenizer from 'src/core/Tokenizer';
 import { EOF_TOKEN, isReserved, isToken, type Token, TokenType } from 'src/core/token';
 import { dedupe } from 'src/utils';
